Moves some directories around
Shoves more stuff inside util as the root directory was getting too bloated, now util is bloated instead.
