+++ backend/package.json
@@ -13,8 +13,7 @@
 		"express": "^4.19.1",
 		"mariadb": "^3.3.0",
 		"openai": "^4.47.1",
-		"sequelize": "^6.37.3",
-		"validator": "^13.12.0"
+		"sequelize": "^6.37.3"
 	},
 	"devDependencies": {
 		"@types/express": "^4.17.21",
