+++ backend/src/index.ts
@@ -1,10 +1,10 @@
 import "./util/environment.js";
 import { getSequelize } from "./util/sequelize.js";
-import { setupControllers } from "./setup/controllers.js";
-import { setupFrontend } from "./setup/frontend.js";
-import { setupHandlers } from "./setup/handlers.js";
-import { setupMiddleware } from "./setup/middleware.js";
-import { setupSwagger } from "./setup/swagger.js";
+import { setupControllers } from "./util/setup/controllers.js";
+import { setupFrontend } from "./util/setup/frontend.js";
+import { setupHandlers } from "./util/setup/handlers.js";
+import { setupMiddleware } from "./util/setup/middleware.js";
+import { setupSwagger } from "./util/setup/swagger.js";
 import chalk from "chalk";
 import express from "express";
 import logging from "./util/logging.js";
